More bignum fixes.
Don't crunch length 1 bignums in LOGfoo.
Deal with the bignum sign changing better in +/-.
The overflow flag is set when the sign changes, and the new sign can be
recovered from the carry flag. If the sign doesn't change then the carry
flag doesn't reflect the new state of the sign bit, so the last fragment
must be sign-extended.