Merge the application and closure classes.
The only permitted object in the BODY slot of a closure was an application
and applications could only show up there, so there wasn't much use in
treating them specially.
Normal calls to symbols are now represented as calls to constants, instead
of directly storing the symbol in the function position. Now the IR
looks a bit like a Lisp-1.