Multiprocessing bugfixes.
Processes now reset their whostate after waking up, clearing the stale
wait reason.
process-wait-with-timeout's timeout check was nonsense.
%maybe-preempt-from-interrupt-frame must avoid calling get-next-process
if it doesn't intend to actually switch to it. get-next-process will
advance to the next runnable process, even if the process it returns
does not get switched to.
idle-process needs to run with interrupts mostly off.