Rework dynamic-extent allocations.
NOTE: All .LLF files must be rebuilt!
This completely solves the ever-expanding-stack problem that
MULTIPLE-VALUE-PROG1 had, makes the GC immune to stray stale DX values
and obsoletes the stack mark bit, which has been removed now.

DX values are now allocated using two different methods, depending on
whether their size is know at compile time or not.
If the size is known, then they will be allocated directly in the
function's stack frame as scannable/live values.
Otherwise, they are allocated on the stack, and a value to the object
tagged with +TAG-DX-ROOT-OBJECT+ will be stored in the stack frame. Only
objects can be allocated this way, not conses. Allocating an arbitrary
number of conses requires them to be wrapped in a simple-vector, possible
because conses are normally represented without any header.

The GC will ignore any values that point to the stack area unless they
have the +TAG-DX-ROOT-OBJECT+ tag. Such values are only permitted to
exist in a function's stack-frame. In other words, they are internal to
the GC and should not be treated as real Lisp values. A real Lisp value
must be tagged with +TAG-OBJECT+ instead. They should remain live for as
long as the dynamic-extent allocation.

Arbitrary-size allocations are not currently exposed by the compiler and
are only used internally for &REST lists and M-V-PROG1's value save area.