Don't write to the disk freelist log in STORE-FREE.
The store freelist log (the on-disk representation) gets fully rebuilt
when a snapshot is taken.
Updating it causes blocks to be added to the block cache, which is
currently never trimmed. Unused cached blocks would build up and slow
down any accesses to the block cache.
This indirectly speeds up anything that accesses the block map, mainly
the allocators & GC.