Fix semaphores and implement new sync objects.
IRQ FIFOs are FIFOs that're safe to push objects into from interrupt
context. Regular FIFOs are currently safe, but at the expense of normal
code.

Latches are like a gate that can be open or closed. When it is closed,
threads will block when LATCH-WAIT is called, when it is open, threads
will not block in LATCH-WAIT (and any blocked threads will be resumed).