Make PANIC more robust.
When using the serial debug stream, immediately switch over to the
early serial stream. This allows PANIC to run with interrupts disabled,
and stops nested panics that were occuring in interrupt handlers.
Bug: This causes any pending serial output to be truncated, losing
some context.
Printing the panic backtrace will catch page faults when the stack is
unmapped. The code has been extracted into a seperate function allowing it
to be called from the pager, to print the stack of whatever function has
faulted.