Fix a possible bug involving DX allocations and NLX.
(lambda ()
  (block nil
    (multiple-value-prog1 (foo)
      (bar (lambda () (return))))))

This code would establish a non-local exit, then create a dynamic-extent
object to hold the values from (FOO). If the lambda was then called
and the block returned from, then M-V-P1's cleanup code wouldn't be run
and the dynamic-extent root that was created would remain live after
the stack pointer was restored by the NLX.
The compiler now tracks when dynamic-extent roots are created and inserts
code to clear them as part of the NLX. The GC assists by ignoring any
DX roots that point below the stack pointer.