Possibly fix the long-standing hash table bug.
I'm still not sure what the problem was, but I think that a garbage
collection was occuring (due to networking or similar, not the hash-table
code) after (SETF GETHASH) had computed the hash, and before
FIND-HASH-TABLE-SLOT ran. This would trigger a rehash, then
FIND-HASH-TABLE-SLOT would search the table with the *stale* hash, leading
to trouble.