Store slot storage layout vectors in each (funcallable-)std-instance.
Preparing for class redefinition. Every instance needs to have a way
of tracking if is obsolete. When a class is redefined, a new layout vector
will be created and every old instance of that class will have a stale
layout cell. The storage layout vector was chosen because
UPDATE-INSTANCE-FOR-REDEFINED-CLASS only needs to know names & values of
the old slots, not the slot-defintion object.

No memory overhead for std-instances, this replaces the padding word.
Funcallable-std-instances have increased in size by two words.