Fix off-by-one error in the symbol binding code.
Symbol TLS slot ids are regular slot indexes, indexing into the thread
code. +TLS-BASE-OFFSET+ was causing symbol accessors to access from the
very start of the thread object, not from the address of the first slot.

This caused TLS accesses to the very first bound symbol, which was
*ACTIVE-CATCH-HANDLERS*, to overwrite the last multiple-values slot.
This slot was always initialized to 0, not to the tls unbound value,
causing the chain of catch tags/handlers to have an invalid terminator.

MAKE-THREAD worked around this bug when setting initial bindings.