Cold generator improvements.
A file-stream can be passed as the image name, and MAKE-IMAGE will
write to that instead of creating a new output file.
The path to the map file can be specified, instead of being derived from
the image name.
An image generated without a header-path or image-size will be as small
as possible. It's assumed that it'll be dd'd onto a real partition.