Run allocate/release/protect-memory-range in the pager thread.
Previously these were run in the context of the calling thread. Normal
threads have pageable stacks, which require the pager to page them in.
If a thread took a stack page fault while in a foo-memory-range function
with the *VM-LOCK* held, the system could deadlock. By running them in
the pager thread, this problem is avoided because the pager is wired.