Do not handle page faults until the paging disk has been found.
If a snapshot is taken when threads in the wait-for-page state, then this
will persist into the next boot. The pager will immediately try to resolve
these page faults during boot, and fail because the block map will not
have been initialised.
Delaying the pager until the paging disk has been found and read fixes
this.