Add a cache for symbol binding cell lookups.
This largely eliminates the slowdown from deep binding, with the cache
having a hit rate of around 90-95%.

The cache is a 128 entry array stored directly in the thread object,
replacing the old TLS array. It is currently index by the low bits of
the symbol address, but this could be replaced by any hash for the symbol.
The cache is updated when there is a miss, when a special variable is
bound (a recently bound special variable will probably be accessed soon),
and when a binding is undone (because it is no longer valid).

This currently relies on the fact that symbols and their value cells
are wired. Changes to the GC and cache index will be needed if
this changes.