Improve how special variables and constants are handled.
This introduces DEFGLOBAL, for declaring a special variable that cannot
be bound over. This avoids a lookup when accessing the symbol's value,
which is important in the supervisor.
Constants can no longer be let-bound over, the compiler signals an error.
There is optional support for SBCL-style defconstant checking, to prevent
constants from being redefined with different values.