Pages can now be marked wired and the block map is loaded by the bootloader
The bootloader must load the block map to determine which pages to load
and where to load them from, so passing it over to the pager prevents
the pager from having to redundantly load it.
Allowing blocks to be marked as wired kills off the extent table and
will pave the way to more flexible use of the address space.
Currently wired blocks are restricted to regions that were previously
wired, but this will change in the future.