Replace the setf fdefinition LLF special case.
This adds an LLF op that calls a known function with arguments. Top-level
function calls are transformed recursively to this, allowing the
cold-generator to inspect which functions are being called.
The file-compiler now only generates opaque functions when a complex
special-operator is encountered at the top-level.

This will allow for the cold-generator to special case more top-level
forms without needing additional logic in LLF handling.