More fixes
Window size accounts for the frame size.
Pause events are signalled with conditions instead of catch/throw.
Video is drawn directly to the framebuffer, avoiding a blit.
Removed AVI argument from play-video-stream's thread lambda.
Changed the event loop to poll the fifo instead of blocking.
This was causing the video player to stop playing unless it was
constantly receiving events.