Fix hash-table iteration.
If a hash-table was being iterated over and accessed through GETHASH
then a GC cycle could cause the iterator to lose track of the current
position. GETHASH after a GC will rehash the hash-table, which creates
a new hash-table storage vector with a potentially different layout.
The iterator's index would not be valid in this new storage vector.

The hash-table iterator has been modified to always use the original
storage vector, not the current.

This was causing the compiler to miscompile code.