Modify probe-file to work correctly on local directories.
Change probe-file to be stand alone instead of using open so that it will work correctly on local directries as well as files.
    
In fs.lisp, probe-file generate an error for pathnames with wildcards then call probe-using-host.
    
In local.lisp implement probe-using-host for the local filesytem.
    
In remote.lisp implement probe-using-host for the remote filesyste.