Fix issue with probe-file.
Change probe-file to use probe-file instead of open :direction :probe because open :direction :probe fails on directories or files that exist but are not accesable due to permissions.

This caused an error in clim-listener in the "Show Directory" command. If the directory contained a file or sub-directory that was inaccessible due to permissions, clim-listener entered the error handler. (For example, "Show Directory" on #P"REMOTE:/" which contains "lost+found" which is only accessible by root.)