Implement CHANGE-CLASS and fix bugs.
The heap layout is now saved when saving structure definitions to avoid
having to reconstruct it at load-time.
The obsolete instance type has been added, this instance uses an
OBSOLETE-INSTANCE-LAYOUT as the header layout instead of a regular
LAYOUT. Accesses to slots should use the new instance instead of the
obsolete instance. This maintains object identity when instances are
made obsolete, but allows the object layout to change.
The GC will rewrite pointers to the old obsolete instance to point at
the new instance to eliminate the indirection overhead.
Some funcallable instances were two words too short, CLOS boot was
not correctly computing the layout for those classes correctly.
Metaobject slot accessors that use a hard-coded location now only do
so on direct instances of their metaobjects, not subclasses. Slot
locations in a superclass may not be valid subclasses.
DEFSTRUCT was not using the short layout representation for fully
unboxed structs.
The file compiler has been updated for the new world.