Translate drive letters in the file-server on Windows hosts.
This is enough to support building the system on Windows.
The source & home directory paths must be set to the form
"/c/path/to/code/", instead of "c:\\path\\to\\code\\"
No other changes are needed, however the MBuild Makefile does not do this
automatically so config.lisp must be edited manually and
build-cold-image.lisp run by hand.