Fix an early crash bug.
Setting the event-state of prior *BOOT-ID* events could cause threads to
be woken up and this would cause a wakeup IPI to be broadcast to other
CPUs. At this point in the boot process the local APIC hasn't been mapped
in, so the IPI functions were touching unmapped memory and causing an
early page fault.

This only occurred in qemu and not in VirtualBox, presumably due to a
difference in memory maps.