Properly save function fixups when file-compiling.
This fixes a bug where completely invalid entries were being written
into the symbol value cell cache and corrupting the system.
The cache sentinel value was different between images and was saved into
the compiled file. Without fixups to update it when the compiled file
was loaded the old and incorrect value was used.