Fix bug in directory-using-host.
The CL function "directory" should not include "." and ".."
directories in the list of files and directories that is returned. At
least, this seems to be the case for other hosts on mezzano (eg LOCAL
and REMOTE) and for directories on linux/SBCL, both ext4 file systems
and fat32 file systems.