Fix buffer leaks when deleting a USB device.
Changed td convention to always have an associated td-info structure
in the td->td-info hash. So, td-info can be allocated at the same time
as the td and the td-info structure can be removed from the
td->td-info hash when the td is freed. This guarantees that the
td->td-info hash is always cleaned up.

Added code for cleaning up bulk transfer eds when a device is
deleted. This lead to creating better primitives for cleaning up eds,
which are now used where appropriate.