Update FAT open-using-host to handle supersede.
Re-organized open-using-host based on the version in local.lisp. So in
layout, the function looks quite different, but functionally it is
only a little different. The major change functionally, was to only
get the file meta-data at the start of the function instead of opening
the file. This allowed for detection of a missing directory and easier
handling of the non-existant file case.