dhcp: Trying to fix automatic lease renewal.
On my network the local router doesn't respond to Mezzano's renewal
requests. I am not sure if it is due to the bridged VM network adaptor
or if the DHCP code is sending malformed requests.

The effect is that the VM's local IP changes over time more than it
would if renewals worked properly.