supervisor: Make the page fault hook per-CPU and more reliable.
It is possible that multiple CPUs may want to establish a hook at the
same time, so the hook variable must be per-CPU, not stored in a global.
It also now provides a safe way to exit the hook via non-local exit
using the ABANDON-PAGE-FAULT local macro. Page fault hooks must not be
exited via normal non-local exits such as RETURN or GO.