Use direct calls when doing ordinary named calls to functions.
Functions and frefs are now stored in a 2GB region, which is within
range of x86's call/jmp instructions.
Frefs have a small chunk of code associated with them that contains
a direct jump to the associated function's entry point and calls via
frefs are done by jumping to this code.

MAKE-FUNCTION-REFERENCE/(SETF FUNCTION-REFERENCE-FUNCTION) have not
been updated, so the system currently crashes during cold start.