hash-table: Minor improvements
Remove the (SETF HASH-TABLE-COUNT) function. H-T-C is a standard CL
function and doesn't have a defined setter.
Check types properly for the rehash-size and -threshold fields.
Restrict the result of HASH-TABLE-SYNCHRONIZED to NIL or T and make it
a read-only field.
Export the getter functions for the synchronized/gc-invariant/weakness
properties.
Print hash-tables more concisely.