Disk: Add support to manage list of all block devices.
Add register-block-device/unregister-block-device/all-block-devices
interface to maintain a list of block devices and partitions. The list
can be used to look for devices or partitions that contain file
systems.

By default, the list contains all devices and partitions that are
supported by the supervisor disk driver (the results of
mezzano.supervisor:all-disk). In addition, the USB mass storage driver
registers and unregisters USB drives and partitions.