compiler: Remove %ARRAY-DIMENSION-KNOWN.
Calls to this internal function could escape the compiler and end up
as real calls to an undefined function when types were too hairy.
Instead use ARRAY-DIMENSION, but force the call to occur with high
speed/low safety settings. This has the same effect, but can always
fall back to calling the standard function.