FILE SYSTEM: Support for file system hosts on local disks after reboot.
Define the file-host-mount-mixin class with a default mount method
so that all file systems that do not depend on local disks are
supported by just including file-host-mount-mixin in their class
defintions.

Add file-host-mount-mixin to all file system host classes.

For the FAT file system, mount and probe-block-device are
appropriately defined so that FAT file system hosts still work when
they have been saved in a snapshot and the system is rebooted.

Reboot for EXT4 hosts is not supported. On reboot, mount returns NIL and the
host is deleted.

Add function that attempts to re-mount all of the file system hosts on
reboot and register the function as an "early" boot-hook.

Add defgeneric for probe-block-device.

Add find-local-block-device which searches for a block device for the
given host with the given UUID.

Add document that describes mezzano file system architecture.