FILE SYSTEM: Changes to FAT short name handling and directory name matching.
Updated read-short-name to use the DIR_NTRes bits (a field in each
directory entry). In more recent versions of the file system two bits
of DIR_NTRes specify that the name and/or extension are lower case
instead of upper case.

Directory name matching was ignoring case because it was using
equalp. Changed to use string= instead.