file system: Update fat and ext filesystem to use the partition label for the host name.
If a UUID to host name mapping is not provided, create-host will
attempt to create a valid host name from the partition label. If a
valid host name is created, the file system is mounted using that
name. If multiple partitions create the same host name, a number is
appended to make the host names unique.