system: Expose cmpxchg16b through a new macro, DCAS.
DCAS takes two places and performs a pair of atomic CAS operations on them.
The places must either be the CAR and CDR of a CONS, or accessors for
slots in a struct that have been marked with the :DCAS-SIBLING slot option.

FIRST/REST are also supported in place of CAR/CDR.