file: Add FILE-PROPERTIES and SET-FILE-PROPERTIES
These functions can be used to get and set arbitrary properties on file,
as supported by the underlying file systems.

The Common Lisp functions FILE-WRITE-DATE and FILE-AUTHOR are now
implemented on top of FILE-PROPERTIES.

FAT and REMOTE only support reading the :WRITE-DATE property.
LOCAL supports any arbitrary property, except LENGTH, which is the length
of the file and cannot be set. :WRITE-DATE and :CREATION-DATE properties
are maintained automatically.